move pydantic response models to interfaces.py
